# Conflicts:
#	.htaccess
#	app/calls/resources/classes/follow_me.php
#	app/devices/app_languages.php
#	resources/templates/provision/escene/e3xx/{$mac}.xml
#	resources/templates/provision/escene/e3xx/{$mac}_extern.xml

+++ app/xml_cdr/resources/classes/xml_cdr.php
@@ -57,7 +57,7 @@ if (!class_exists('xml_cdr')) {
 		 * unset the variables used in the class
 		 */
 		public function __destruct() {
-			foreach ($this as $key => $value) {
+			if (isset($this)) foreach ($this as $key => $value) {
 				unset($this->$key);
 			}
 		}
@@ -94,6 +94,7 @@ if (!class_exists('xml_cdr')) {
 			$this->fields[] = "caller_id_name";
 			$this->fields[] = "caller_id_number";
 			$this->fields[] = "destination_number";
+			$this->fields[] = "source_number";
 			$this->fields[] = "start_epoch";
 			$this->fields[] = "start_stamp";
 			$this->fields[] = "answer_stamp";
@@ -145,7 +146,7 @@ if (!class_exists('xml_cdr')) {
 
 			$sql = "insert into v_xml_cdr (";
 			$f = 1;
-			foreach ($this->fields as $field) {
+			if (isset($this->fields)) foreach ($this->fields as $field) {
 				if ($field_count == $f) {
 					$sql .= "$field ";
 				}
@@ -159,10 +160,10 @@ if (!class_exists('xml_cdr')) {
 			$row_count = sizeof($this->array);
 			//$field_count = sizeof($this->fields);
 			$i = 0;
-			foreach ($this->array as $row) {
+			if (isset($this->array)) foreach ($this->array as $row) {
 				$sql .= "(";
 				$f = 1;
-				foreach ($this->fields as $field) {
+				if (isset($this->fields)) foreach ($this->fields as $field) {
 					if (isset($row[$field]) && strlen($row[$field]) > 0) {
 						$sql .= "'".$row[$field]."'";
 					}
@@ -217,9 +218,38 @@ if (!class_exists('xml_cdr')) {
 					//$this->log("\nfail loadxml: " . $e->getMessage() . "\n");
 				}
 
+			//get the destination number
+				if ($xml->variables->current_application == "bridge") {
+					$current_application_data = urldecode($xml->variables->current_application_data);
+					$bridge_array = explode("/", $current_application_data);
+					$destination_number = end($bridge_array);
+					if (strpos($destination_number,'@') !== FALSE) {
+						$destination_array = explode("@", $destination_number);
+						$destination_number = $destination_array[0];
+					}
+				}
+				else {
+					$destination_number = urldecode($xml->variables->sip_to_user);
+				}
+
+			//get the caller id
+				$caller_id_name = urldecode($xml->variables->effective_caller_id_name);
+				$caller_id_number = urldecode($xml->variables->effective_caller_id_number);
+				if (strlen($caller_id_number) == 0) foreach ($xml->callflow as $row) {
+					$caller_id_name = urldecode($row->caller_profile->caller_id_name);
+					$caller_id_number = urldecode($row->caller_profile->caller_id_number);
+				}
+
 			//misc
 				$uuid = check_str(urldecode($xml->variables->uuid));
 				$this->array[$row]['uuid'] = $uuid;
+				$this->array[$row]['destination_number'] = check_str($destination_number);
+				$this->array[$row]['source_number'] = check_str(urldecode($xml->variables->effective_caller_id_number));
+				$this->array[$row]['user_context'] = check_str(urldecode($xml->variables->user_context));
+				$this->array[$row]['network_addr'] = check_str(urldecode($xml->variables->sip_network_ip));
+				$this->array[$row]['caller_id_name'] = check_str($caller_id_name);
+				$this->array[$row]['caller_id_number'] = check_str($caller_id_number);
+
 				$this->array[$row]['accountcode'] = check_str(urldecode($xml->variables->accountcode));
 				$this->array[$row]['default_language'] = check_str(urldecode($xml->variables->default_language));
 				$this->array[$row]['bridge_uuid'] = check_str(urldecode($xml->variables->bridge_uuid));
@@ -268,21 +298,6 @@ if (!class_exists('xml_cdr')) {
 					$this->array[$row]['rtp_audio_in_mos'] = $rtp_audio_in_mos;
 				}
 
-			//get the values from the callflow.
-				$x = 0;
-				foreach ($xml->callflow as $callflow) {
-					if ($x == 0) {
-						$context = check_str(urldecode($callflow->caller_profile->context));
-						$this->array[$row]['destination_number'] = check_str(urldecode($callflow->caller_profile->destination_number));
-						$this->array[$row]['context'] = $context;
-						$this->array[$row]['network_addr'] = check_str(urldecode($callflow->caller_profile->network_addr));
-					}
-					$this->array[$row]['caller_id_name'] = check_str(urldecode($callflow->caller_profile->caller_id_name));
-					$this->array[$row]['caller_id_number'] = check_str(urldecode($callflow->caller_profile->caller_id_number));
-					$x++;
-				}
-				unset($x);
-
 			//store the call leg
 				$this->array[$row]['leg'] = $leg;
 
@@ -325,7 +340,6 @@ if (!class_exists('xml_cdr')) {
 
 			//set values in the database
 				if (strlen($domain_uuid) > 0) {
-					$database->domain_uuid = $domain_uuid;
 					$this->array[$row]['domain_uuid'] = $domain_uuid;
 				}
 				if (strlen($domain_name) > 0) {
@@ -432,7 +446,6 @@ if (!class_exists('xml_cdr')) {
 			$xml_cdr_dir = $_SESSION['switch']['log']['dir'].'/xml_cdr';
 			$dir_handle = opendir($xml_cdr_dir);
 			$x = 0;
-			$cdr = new $xml_cdr;
 			while($file = readdir($dir_handle)) {
 				if ($file != '.' && $file != '..') {
 					if ( !is_dir($xml_cdr_dir . '/' . $file) ) {
@@ -446,21 +459,53 @@ if (!class_exists('xml_cdr')) {
 								$file_prefix = substr($file, 0, 1);
 							}
 
-						//get the xml cdr string
-							$xml_string = file_get_contents($xml_cdr_dir.'/'.$file);
+						//set the limit
+							if (isset($_SERVER["argv"][1]) && is_numeric($_SERVER["argv"][1])) {
+								$limit = $_SERVER["argv"][1];
+							}
+							else {
+								$limit = 1;
+							}
 
-						//parse the xml and insert the data into the db
-							$cdr->xml_array($x, $leg, $xml_string);
+						//filter for specific files based on the file prefix
+							if (isset($_SERVER["argv"][2])) {
+								if (strpos($_SERVER["argv"][2], $file_prefix) !== FALSE) {
+									$import = true;
+								}
+								else {
+									$import = false;
+								}
+							}
+							else {
+								$import = true;
+							}
 
-						//delete the file after it has been imported
-							unlink($xml_cdr_dir.'/'.$file);
+						//import the call detail record
+							if ($import) {
+								//get the xml cdr string
+									$xml_string = file_get_contents($xml_cdr_dir.'/'.$file);
 
-						//increment
-							$x++;
+								//parse the xml and insert the data into the db
+									$this->xml_array($x, $leg, $xml_string);
+
+								//delete the file after it has been imported
+									unlink($xml_cdr_dir.'/'.$file);
+							}
+
+						//increment the value
+							if ($import) {
+								$x++;
+							}
+
+						//if limit exceeded exit the loop
+							if ($limit == $x) {
+								//echo "limit: $limit count: $x if\n";
+								break;
+							}
 					}
 				}
 			}
-			$cdr->save();
+			$this->save();
 			closedir($dir_handle);
 		}
 		//$this->read_files();
@@ -488,7 +533,7 @@ if (!class_exists('xml_cdr')) {
 								catch(Exception $e) {
 									echo $e->getMessage();
 								}
-								foreach ($conf_xml->settings->param as $row) {
+								if (isset($conf_xml->settings->param)) foreach ($conf_xml->settings->param as $row) {
 									if ($row->attributes()->name == "cred") {
 										$auth_array = explode(":", $row->attributes()->value);
 										//echo "username: ".$auth_array[0]."<br />\n";
@@ -544,9 +589,8 @@ if (!class_exists('xml_cdr')) {
 					//xml_cdr_log("process cdr via post\n");
 
 				//parse the xml and insert the data into the database
-					$cdr = new $xml_cdr;
-					$cdr->xml_array(0, $leg, $xml_string);
-					$cdr->save();
+					$this->xml_array(0, $leg, $xml_string);
+					$this->save();
 			}
 		}
 		//$this->post();
@@ -556,6 +600,6 @@ if (!class_exists('xml_cdr')) {
 /*
 //example use
 	$cdr = new xml_cdr;
-	$csv -> $cdr->extension_summary('csv');
+	$cdr->read_files();
 */
 ?>
